Refactors all remaining API calls on the front-end.
