+++ react-native-customer-frontend/businesses/networking/controllers/attendantController.tsx
@@ -1,7 +1,7 @@
-import { attendantsRoute } from "../routers";
+import { attendantsRoute, attendantRoute } from "../routers/attendantRouter";
 
 async function getAttendant(attendantId: string) {
-    return await fetch(attendantsRoute + attendantId, { method: "GET" })
+    return await fetch(attendantRoute(attendantId), { method: "GET" })
         .then(res => res.json());
 }
