+++ react-native-customer-frontend/businesses/networking/routers/appointmentRouter.tsx
@@ -1,0 +1,9 @@
+import { BASE } from ".";
+
+const appointmentsRoute = BASE + "appointments/";
+
+const appointmentRoute = (appointmentId: string) => {
+    return appointmentsRoute + appointmentId;
+}
+
+export { appointmentsRoute, appointmentRoute };
