+++ react-native-customer-frontend/businesses/networking/routers/attendantRouter.tsx
@@ -1,0 +1,9 @@
+import { BASE } from ".";
+
+const attendantsRoute = BASE + "attendants/";
+
+const attendantRoute = (attendantId: string) => {
+    return attendantsRoute + attendantId;
+}
+
+export { attendantsRoute, attendantRoute };
