+++ react-native-customer-frontend/businesses/networking/routers/businessRouter.tsx
@@ -1,0 +1,9 @@
+import { BASE } from ".";
+
+const businessesRoute = BASE + "businesses/";
+
+const businessRoute = (businessId: string) => {
+    return businessesRoute + businessId;
+}
+
+export { businessesRoute, businessRoute };
